move getSysInfo functions and create new local_utils.py
